# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Swagger representation of ScoreTrak's Mater API"
  version: "1.0.0"
  title: "ScoreTrak Mater API"
...
schemes:
  - "https"
  - "http"
paths:
...
  /score:
    get:
      tags:
        - "Score"
      summary: "Retreive all Scores"
      description: "Orders scoring engine to retrive all scores"
      operationId: "GetScores"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/score"
        "401":
          description: "Authentication token is missing, or incorrect"
          headers:
            WWW_Authenticate:
              type: "string"
      security:
        - api_key: []
  /score/{teamID}:
    get:
      tags:
        - "Score"
      summary: "Retreive all Scores for a given team"
      description: "Orders scoring engine to retrive all scores"
      operationId: "GetScore"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "teamID"
          in: "path"
          description: "ID of Host to return"
          required: true
          type: "integer"
          format: "int64"
          x-exportParamName: "TeamID"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/score"
        "401":
          description: "Authentication token is missing, or incorrect"
          headers:
            WWW_Authenticate:
              type: "string"
      security:
        - api_key: []
  /round:
    get:
      tags:
        - "Round"
      summary: "Retreive total number of rounds"
      description: "Orders scoring engine to return total number of rounds"
      operationId: "GetLastRound"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/round"
        "401":
          description: "Authentication token is missing, or incorrect"
          headers:
            WWW_Authenticate:
              type: "string"
      security:
        - api_key: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  score:
    type: "object"
    properties:
      team_id:
        type: "integer"
        format: "int64"
        example: 20
      score:
        type: "integer"
        format: "int64"
        example: 1000
    description: "Score object shows scores for a given team"
    example:
      score: 1000
      team_id: 20
...
  round:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Round ID can also represent the round number"
      start:
        type: "string"
        format: "date-time"
      end:
        type: "string"
        format: "date-time"
    description: "Round Model holds the information about all the rounds that have\
      \ passed"
    example:
      start: "2000-01-23T04:56:07.000+00:00"
      end: "2000-01-23T04:56:07.000+00:00"
      id: 0
      passed: true
...